PR-AN-06: Refactor - Phase 2
