+++ lib/api/command.js
@@ -1,0 +1,9 @@
+const argsert = require('../argsert')
+
+module.exports = function (context, command) {
+  return function (cmd, description, builder, handler, middlewares) {
+    argsert('<string|array|object> [string|boolean] [function|object] [function] [array]', [cmd, description, builder, handler, middlewares], arguments.length)
+    command.addHandler(cmd, description, builder, handler, middlewares)
+    return context
+  }
+}
