+++ lib/api/config.js
@@ -1,0 +1,27 @@
+const argsert = require('../argsert')
+
+module.exports = function (options, context, cwd, usage, applyExtends) {
+  return function config (key, msg, parseFn) {
+    argsert('[object|string] [string|function] [function]', [key, msg, parseFn], arguments.length)
+    // allow a config object to be provided directly.
+    if (typeof key === 'object') {
+      key = applyExtends(key, cwd)
+      options.configObjects = (options.configObjects || []).concat(key)
+      return context
+    }
+
+    // allow for a custom parsing function.
+    if (typeof msg === 'function') {
+      parseFn = msg
+      msg = null
+    }
+
+    key = key || 'config'
+    context.describe(key, msg || usage.deferY18nLookup('Path to JSON config file'))
+    ;(Array.isArray(key) ? key : [key]).forEach((k) => {
+      options.config[k] = parseFn || true
+    })
+
+    return context
+  }
+}
