+++ lib/api/usage.js
@@ -1,0 +1,21 @@
+const YError = require('../yerror')
+const argsert = require('../argsert')
+
+module.exports = function (options, context, usage) {
+  return function (msg, description, builder, handler) {
+    argsert('<string|null|undefined> [string|boolean] [function|object] [function]', [msg, description, builder, handler], arguments.length)
+
+    if (description !== undefined) {
+      // .usage() can be used as an alias for defining
+      // a default command.
+      if ((msg || '').match(/^\$0( |$)/)) {
+        return context.command(msg, description, builder, handler)
+      } else {
+        throw new YError('.usage() description must start with $0 if being used as alias for .command()')
+      }
+    } else {
+      usage.usage(msg)
+      return context
+    }
+  }
+}
