+++ test/yargs.spec.js
@@ -2,11 +2,9 @@
 /* global context, describe, it, beforeEach, afterEach */
 
 const expect = require('chai').expect
-const fs = require('fs')
 const path = require('path')
 const checkOutput = require('./helpers/utils').checkOutput
 let yargs
-const YError = require('../lib/yerror')
 
 require('chai').should()
 
@@ -306,224 +304,6 @@ describe('yargs dsl tests', () => {
         expect(err).to.equal(null)
         return done()
       })
-    })
-  })
-
-  describe('command', () => {
-    it('executes command handler with parsed argv', (done) => {
-      yargs(['blerg'])
-        .command(
-          'blerg',
-          'handle blerg things',
-          noop,
-          (argv) => {
-            // we should get the argv from the prior yargs.
-            argv._[0].should.equal('blerg')
-            return done()
-          }
-        )
-        .exitProcess(false) // defaults to true.
-        .argv
-    })
-    it('runs all middleware before reaching the handler', function (done) {
-      yargs(['foo'])
-        .command(
-          'foo',
-          'handle foo things',
-          function () {},
-          function (argv) {
-            // we should get the argv filled with data from the middleware
-            argv._[0].should.equal('foo')
-            argv.hello.should.equal('world')
-            return done()
-          },
-        [function (argv) {
-          return {hello: 'world'}
-        }]
-        )
-        .exitProcess(false) // defaults to true.
-        .argv
-    })
-    it('recommends a similar command if no command handler is found', () => {
-      const r = checkOutput(() => {
-        yargs(['boat'])
-          .command('goat')
-          .recommendCommands()
-          .argv
-      })
-
-      r.errors[2].should.match(/Did you mean goat/)
-    })
-
-    it('does not recommend a similiar command if no similar command exists', () => {
-      const r = checkOutput(() => {
-        yargs(['foo'])
-          .command('nothingSimilar')
-          .recommendCommands()
-          .argv
-      })
-
-      r.logs.should.be.empty
-    })
-
-    it('recommends the longest match first', () => {
-      const r = checkOutput(() => {
-        yargs(['boat'])
-          .command('bot')
-          .command('goat')
-          .recommendCommands()
-          .argv
-      })
-
-      r.errors[2].should.match(/Did you mean goat/)
-    })
-
-    // see: https://github.com/yargs/yargs/issues/822
-    it('does not print command recommendation if help message will be shown', (done) => {
-      const parser = yargs()
-        .command('goat')
-        .help()
-        .recommendCommands()
-
-      parser.parse('boat help', {}, (err, _argv, output) => {
-        if (err) return done(err)
-        output.split('Commands:').length.should.equal(2)
-        return done()
-      })
-    })
-
-    it("skips executing root-level command if builder's help is executed", () => {
-      const r = checkOutput(() => {
-        yargs(['blerg', '-h'])
-          .command(
-            'blerg',
-            'handle blerg things',
-            yargs => yargs
-                .command('snuh', 'snuh command')
-                .help('h')
-                .wrap(null),
-            () => {
-              throw Error('should not happen')
-            }
-          )
-          .help('h')
-          .argv
-      })
-
-      r.logs[0].split('\n').should.deep.equal([
-        'usage blerg',
-        '',
-        'handle blerg things',
-        '',
-        'Commands:',
-        '  usage blerg snuh  snuh command',
-        '',
-        'Options:',
-        '  --version  Show version number  [boolean]',
-        '  -h         Show help  [boolean]'
-      ])
-    })
-
-    it('executes top-level help if no handled command is provided', () => {
-      const r = checkOutput(() => {
-        yargs(['snuh', '-h'])
-          .command('blerg', 'handle blerg things', yargs => yargs
-              .command('snuh', 'snuh command')
-              .help('h')
-              .argv
-            )
-          .help('h')
-          .wrap(null)
-          .argv
-      })
-
-      r.logs[0].split('\n').should.deep.equal([
-        'usage [command]',
-        '',
-        'Commands:',
-        '  usage blerg  handle blerg things',
-        '',
-        'Options:',
-        '  --version  Show version number  [boolean]',
-        '  -h         Show help  [boolean]'
-      ])
-    })
-
-    it("accepts an object for describing a command's options", () => {
-      const r = checkOutput(() => {
-        yargs(['blerg', '-h'])
-          .command('blerg <foo>', 'handle blerg things', {
-            foo: {
-              default: 99
-            },
-            bar: {
-              default: 'hello world'
-            }
-          })
-          .help('h')
-          .wrap(null)
-          .argv
-      })
-
-      const usageString = r.logs[0]
-      usageString.should.match(/usage blerg <foo>/)
-      usageString.should.match(/--foo.*default: 99/)
-      usageString.should.match(/--bar.*default: "hello world"/)
-    })
-
-    it("accepts a module with a 'builder' and 'handler' key", () => {
-      const argv = yargs(['blerg', 'bar'])
-        .command('blerg <foo>', 'handle blerg things', require('./fixtures/command'))
-        .argv
-
-      argv.banana.should.equal('cool')
-      argv.batman.should.equal('sad')
-      argv.foo.should.equal('bar')
-
-      global.commandHandlerCalledWith.banana.should.equal('cool')
-      global.commandHandlerCalledWith.batman.should.equal('sad')
-      global.commandHandlerCalledWith.foo.should.equal('bar')
-      delete global.commandHandlerCalledWith
-    })
-
-    it("accepts a module with a keys 'command', 'describe', 'builder', and 'handler'", () => {
-      const argv = yargs(['blerg', 'bar'])
-        .command(require('./fixtures/command-module'))
-        .argv
-
-      argv.banana.should.equal('cool')
-      argv.batman.should.equal('sad')
-      argv.foo.should.equal('bar')
-
-      global.commandHandlerCalledWith.banana.should.equal('cool')
-      global.commandHandlerCalledWith.batman.should.equal('sad')
-      global.commandHandlerCalledWith.foo.should.equal('bar')
-      delete global.commandHandlerCalledWith
-    })
-
-    it('derives \'command\' string from filename when missing', () => {
-      const argv = yargs('nameless --foo bar')
-        .command(require('./fixtures/cmddir_noname/nameless'))
-        .argv
-
-      argv.banana.should.equal('cool')
-      argv.batman.should.equal('sad')
-      argv.foo.should.equal('bar')
-
-      global.commandHandlerCalledWith.banana.should.equal('cool')
-      global.commandHandlerCalledWith.batman.should.equal('sad')
-      global.commandHandlerCalledWith.foo.should.equal('bar')
-      delete global.commandHandlerCalledWith
-    })
-
-    it('throws error for non-module command object missing \'command\' string', () => {
-      expect(() => {
-        yargs.command({
-          desc: 'A command with no name',
-          builder (yargs) { return yargs },
-          handler (argv) {}
-        })
-      }).to.throw(/No command name given for module: { desc: 'A command with no name',\n {2}builder: \[Function(: builder)?],\n {2}handler: \[Function(: handler)?] }/)
     })
   })
 
@@ -1147,123 +927,6 @@ describe('yargs dsl tests', () => {
           '',
           'Must call a command'
         ])
-      })
-    })
-  })
-
-  describe('config', () => {
-    it('allows a parsing function to be provided as a second argument', () => {
-      const argv = yargs('--config ./test/fixtures/config.json')
-        .config('config', cfgPath => JSON.parse(fs.readFileSync(cfgPath)))
-        .global('config', false)
-        .argv
-
-      argv.foo.should.equal('baz')
-    })
-
-    it('allows key to be specified with option shorthand', () => {
-      const argv = yargs('--config ./test/fixtures/config.json')
-        .option('config', {
-          config: true,
-          global: false
-        })
-        .argv
-
-      argv.foo.should.equal('baz')
-    })
-
-    it('can be disabled with option shorthand', () => {
-      const argv = yargs('--config ./test/fixtures/config.json')
-        .option('config', {
-          config: false,
-          global: false
-        })
-        .argv
-
-      argv.config.should.equal('./test/fixtures/config.json')
-    })
-
-    it('allows to pass a configuration object', () => {
-      const argv = yargs
-          .config({foo: 1, bar: 2})
-          .argv
-
-      argv.foo.should.equal(1)
-      argv.bar.should.equal(2)
-    })
-
-    describe('extends', () => {
-      it('applies default configurations when given config object', () => {
-        const argv = yargs
-          .config({
-            extends: './test/fixtures/extends/config_1.json',
-            a: 1
-          })
-          .argv
-
-        argv.a.should.equal(1)
-        argv.b.should.equal(22)
-        argv.z.should.equal(15)
-      })
-
-      it('protects against circular extended configurations', () => {
-        expect(() => {
-          yargs.config({extends: './test/fixtures/extends/circular_1.json'})
-        }).to.throw(YError)
-      })
-
-      it('handles aboslute paths', () => {
-        const absolutePath = path.join(process.cwd(), 'test', 'fixtures', 'extends', 'config_1.json')
-
-        const argv = yargs
-          .config({
-            a: 2,
-            extends: absolutePath
-          })
-          .argv
-
-        argv.a.should.equal(2)
-        argv.b.should.equal(22)
-        argv.z.should.equal(15)
-      })
-
-      // see: https://www.npmjs.com/package/yargs-test-extends
-      it('allows a module to be extended, rather than a JSON file', () => {
-        const argv = yargs()
-          .config({
-            a: 2,
-            extends: 'yargs-test-extends'
-          })
-          .argv
-
-        argv.a.should.equal(2)
-        argv.c.should.equal(201)
-      })
-
-      it('ignores an extends key that does not look like a path or module', () => {
-        const argv = yargs()
-          .config({
-            a: 2,
-            extends: 'batman'
-          })
-          .argv
-
-        argv.a.should.equal(2)
-        argv.extends.should.equal('batman')
-      })
-
-      it('allows files with .*rc extension to be extended', () => {
-        const argv = yargs()
-          .config({
-            extends: './test/fixtures/extends/.myrc',
-            a: 3
-          })
-          .argv
-
-        argv.a.should.equal(3)
-        argv.b.should.equal(22)
-        argv.c.should.equal(201)
-        argv.z.should.equal(15)
       })
     })
   })
